+++ src/components.d.ts
@@ -775,14 +775,6 @@ export namespace Components {
     }
     interface SlMenu {
         /**
-          * Removes focus from the menu.
-         */
-        "removeFocus": () => Promise<void>;
-        /**
-          * Sets focus on the menu.
-         */
-        "setFocus": () => Promise<void>;
-        /**
           * Initiates type-to-select logic, which automatically selects an option based on what the user is currently typing. The key passed will be appended to the internal query and the selection will be updated. After a brief period, the internal query is cleared automatically. This method is intended to be used with the keydown event. Useful for enabling type-to-select when the menu doesn't have focus.
          */
         "typeToSelect": (key: string) => Promise<void>;
@@ -2468,14 +2460,6 @@ declare namespace LocalJSX {
         "value"?: string;
     }
     interface SlMenu {
-        /**
-          * Emitted when the menu loses focus.
-         */
-        "onSl-blur"?: (event: CustomEvent<any>) => void;
-        /**
-          * Emitted when the menu gains focus.
-         */
-        "onSl-focus"?: (event: CustomEvent<any>) => void;
         /**
           * Emitted when a menu item is selected.
          */